move more stuff to utils and agreement files
